+++ autokuma/src/main.rs
@@ -1,4 +1,5 @@
 use ::config::{Config, Environment, File, FileFormat};
+use flexi_logger::{Cleanup, Criterion, Duplicate, FileSpec, Logger, LoggerHandle, Naming};
 use serde_json::json;
 use std::sync::Arc;
 use util::{ResultLogger, ResultOrDie};
@@ -47,37 +48,53 @@ const BANNER: &str = r"
           /_/    \_\ \__,_| \__| \___/ |_|\_\ \__,_||_| |_| |_| \__,_|                                                    
 ";
 
+fn create_logger(config: &Arc<crate::config::Config>) -> LoggerHandle {
+  let mut builder = Logger::try_with_env_or_str("info").unwrap();
+
+  if let Some(log_dir) = config.log_dir.as_ref()
+  {
+    builder = builder
+      .log_to_file(FileSpec::default().directory(log_dir))
+      .append()
+      .rotate(Criterion::Size(1_000_000), Naming::NumbersDirect, Cleanup::KeepLogAndCompressedFiles(1, 5))
+      .duplicate_to_stderr(Duplicate::All);
+
+  } 
+
+  return builder.start().unwrap();
+}
+
 #[tokio::main()]
 async fn main() {
-    pretty_env_logger::formatted_timed_builder()
-        .filter(None, log::LevelFilter::Info)
-        .parse_default_env()
-        .init();
-
-    println!("{}", BANNER);
-
     let config: Arc<crate::config::Config> = Arc::new(
         Config::builder()
             .add_source(File::from_str(
                 &serde_json::to_string(&json!({"kuma": {}, "docker": {}})).unwrap(),
                 FileFormat::Json,
             ))
-            .add_source(File::new("config", FileFormat::Toml).required(false))
+            .add_source(File::with_name(&dirs::config_local_dir().map(|dir| dir.join("autokuma").join("config").to_string_lossy().to_string()).unwrap_or_default()).required(false))
+            .add_source(File::new("autokuma", FileFormat::Toml).required(false))
             .add_source(
                 Environment::with_prefix("AUTOKUMA")
                     .separator("__")
                     .prefix_separator("__"),
             )
             .build()
-            .log_error(std::module_path!(), |e| format!("Unable to load config: {}", e))
+            .print_error(|e| format!("Unable to load config: {}", e))
             .and_then(|config| config.try_deserialize())
-            .log_error(std::module_path!(), |e| format!("Invalid config: {}", e))
+            .print_error(|e| format!("Invalid config: {}", e))
             .unwrap_or_die(1),
     );
+  
+    let logger = create_logger(&config);
+
+    println!("{}", BANNER);
 
     let sync = sync::Sync::new(config)
         .log_error(std::module_path!(), |e| format!("Invalid config: {}", e))
         .unwrap_or_die(1);
 
     sync.run().await;
+
+    logger.shutdown();
 }