+++ kuma-client/src/util.rs
@@ -6,6 +6,7 @@ pub trait ResultLogger<F> {
     fn log_info(self, target: &str, cb: F) -> Self;
     fn log_warn(self, target: &str, cb: F) -> Self;
     fn log_error(self, target: &str, cb: F) -> Self;
+    fn print_error(self, cb: F) -> Self;
 }
 
 impl<F, S, T, E> ResultLogger<F> for std::result::Result<T, E>
@@ -47,6 +48,13 @@ where
             e
         });
     }
+
+    fn print_error(self, cb: F) -> Self {
+        return self.map_err(|e| {
+            println!("{}", cb(&e).as_ref());
+            e
+        });
+    }
 }
 
 impl<F, S, T> ResultLogger<F> for Option<T>
@@ -85,6 +93,13 @@ where
     fn log_error(self, target: &str, cb: F) -> Self {
         if self.is_none() {
             error!(target: target, "{}", cb().as_ref())
+        }
+        self
+    }
+
+    fn print_error(self, cb: F) -> Self {
+        if self.is_none() {
+            println!("{}", cb().as_ref())
         }
         self
     }